LVN-40: One single Task Column component instead of the 3 separate ones (ToDo, In Progress, Done)
