--- src/components/Dashboard/TasksColumns/Done.jsx
@@ -1,34 +1,0 @@
-import React, { Component } from 'react';
-import { Container } from 'react-bootstrap';
-import './TaskColumns.scss';
-import TaskItem from '../TaskItem/TaskItem';
-
-export default class Done extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
-
-  render() {
-    const { sortedTasks, taskListRef } = this.props;
-
-    const tasksToDisply = sortedTasks.map(
-      task => (
-        <TaskItem
-          key={task.id}
-          status={task.status}
-          id={task.id}
-          taskName={task.name}
-          taskListRef={taskListRef}
-        />
-      ),
-    );
-
-    return (
-      <Container>
-        <h1 className="taskColumnDoneTitle">Done</h1>
-        <Container className="taskItemsContainer">{tasksToDisply}</Container>
-      </Container>
-    );
-  }
-}
